+++ app/assets/javascript/password.js
@@ -8,16 +8,32 @@ $(document).ready(function () {
             {"data": "account"},
             {"data": "login"},
             {"data": "password"},
-            {"data": "webSite"}/*,
-             { "data": "notes" }*/
+            {"data": "webSite"},
+            {"data": "comments"}
+        ]
+    });
+
+
+    new $.fn.dataTable.Buttons(table, {
+        buttons: [
+            {
+                className: 'fa fa-plus', text: ' Add Password',
+                action: function (e, dt, node, config) {
+                    window.location.replace("/play-pass/password/add");
+                }
+            }
         ]
     });
 
     new $.fn.dataTable.Buttons(table, {
         buttons: [
-            'copy', 'excel', 'pdf', 'print'
+            {extend: 'copy', className: 'fa fa-files-o', text: '', "sToolTip": 'Copy to clipBoard'},
+            {extend: 'excel', className: 'fa fa-file-excel-o', text: ''},
+            {extend: 'pdf', className: 'fa fa-file-pdf-o', text: ''},
+            {extend: 'print', className: 'fa fa-print', text: ''}
         ]
     });
+
 
     table.buttons().container()
         .insertBefore($('.col-sm-12', table.table().container()));