+++ CHANGELOG.md
@@ -3,6 +3,10 @@ v1.0.0 (2016-0X-XX)
 
 * Add/Edit/Delete passwords.
  * change or edit passwords with a helping password generator.
+ 
+* Export options include copy to clipboard, CSV, Excel (XSLX) and PDF files
+
+* Print preview
 
 v0.0.1-SNAPSHOT (2016-02-09) ✓
 ------------------------