+++ CHANGELOG.md
@@ -8,7 +8,7 @@ v1.0.0 (2016-0X-XX)
 * server side user-password hashing with [Heimdall - Secure Password Hashing](https://github.com/qaware/heimdall) ✓
 
 * Add/Edit/Delete passwords.
- * change or edit passwords with a helping password generator.
+ * add or edit passwords with a helping password generator.
 
 v0.0.1-SNAPSHOT (2016-02-09) ✓
 ------------------------